% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_word_cloud.R
\name{clean_tweets}
\alias{clean_tweets}
\title{Create a clean corpus of all the used words in a users tweets.}
\usage{
clean_tweets(df)
}
\arguments{
\item{df}{A dataframe containing a users tweets that has been preprocessed.}
}
\value{
A VCorpus of all the words used disregarding stopwords and punctuation.
}
\description{
Create a clean corpus of all the used words in a users tweets.
}
\examples{
\dontrun{
  docs <- clean_tweets(df)
}
}
