% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{count_tweets}
\alias{count_tweets}
\title{Count the propotion of different sentiment tweets in a labelled sentiment dataframe}
\usage{
count_tweets(df, proportion = TRUE)
}
\arguments{
\item{df}{A dataframe after pre-processing}

\item{proportion}{A boolean value if True: returns the proportion; otherwise, return the counts}
}
\value{
A dictionary which calculates the proportion of three sentiments of tweets
}
\description{
Count the propotion of different sentiment tweets in a labelled sentiment dataframe
}
\examples{
x <- count_tweets(labelled_df)
}
