% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_word_cloud.R
\name{count_words}
\alias{count_words}
\title{Create a dataframe with all the words in a users tweets and how often they appear.}
\usage{
count_words(docs)
}
\arguments{
\item{docs}{A VCorpus of words used in a users tweets.}
}
\value{
A dataframe with all the words in the tweets and how often they appear.
}
\description{
Create a dataframe with all the words in a users tweets and how often they appear.
}
\examples{
\dontrun{
  df2 <- count_words(docs)
}
}
