% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_word_cloud.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{Create a wordcloud of a users tweets.}
\usage{
create_wordcloud(df)
}
\arguments{
\item{df}{A dataframe with all the words in the tweets and how often they appear.}
}
\value{
A wordcloud of a users tweets.
}
\description{
Create a wordcloud of a users tweets.
}
\examples{
\dontrun{
  create_wordcloud(df2)
}
}
