% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{generalPreprocessing}
\alias{generalPreprocessing}
\title{Perform general preprocessing on df. Removes retweets/favourites and cleans URLs, Mentions, and Numbers.}
\usage{
generalPreprocessing(df)
}
\arguments{
\item{df}{dataframe storing all the raw data with text column}
}
\value{
processed tweet dataframe.
}
\description{
Perform general preprocessing on df. Removes retweets/favourites and cleans URLs, Mentions, and Numbers.
}
\examples{
df <- data.frame(User = c("James"), text = c("Hi, how are you!"))
x <- generalPreprocessing(df)
}
