% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_twitter_msg.R
\name{load_twitter_by_keywords}
\alias{load_twitter_by_keywords}
\title{Load dataframe which contains specific user and return as a dataframe with total tweets}
\usage{
load_twitter_by_keywords(user_name, limit, user_info)
}
\arguments{
\item{limit}{An integer shows how many tweets user want to return}

\item{user}{A str contains user_name}
}
\value{
A dataframe
}
\description{
Load dataframe which contains specific user and return as a dataframe with total tweets
}
\examples{
\dontrun{
x <- load_twitter_by_user('nice', 50, user)
}
}
