% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_twitter_msg.R
\name{load_twitter_by_user}
\alias{load_twitter_by_user}
\title{Load dataframe which contains specific user and return as a dataframe with total tweets}
\usage{
load_twitter_by_user(user_name, limit, user_info)
}
\arguments{
\item{user_name}{A str contains tweets with specific user}

\item{limit}{An integer shows how many tweets user want to return}

\item{user_info}{A list contains user info}
}
\value{
A dataframe contains all tweets contains specific user tweets
}
\description{
Load dataframe which contains specific user and return as a dataframe with total tweets
}
\examples{
\dontrun{
x <- load_twitter_by_user('Cristiano', 100, user1)
}
}
