% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_twitter_msg.R
\name{user_info}
\alias{user_info}
\title{Title Take four input parameters and stored in a list object
named "user_info" with keys "consumer_key", "consumer_secret", "access_token", and "access_token_secret".}
\usage{
user_info(
  consumer_key_input,
  consumer_secret_input,
  access_token_input,
  access_token_secret_input
)
}
\arguments{
\item{consumer_key_input}{twitter consumer key}

\item{consumer_secret_input}{twitter consumer secret}

\item{access_token_input}{twitter token}

\item{access_token_secret_input}{twitter token secret}
}
\value{
list contains above user info
}
\description{
Title Take four input parameters and stored in a list object
named "user_info" with keys "consumer_key", "consumer_secret", "access_token", and "access_token_secret".
}
\examples{
\dontrun{
user_info(consumer_key_input, consumer_secret_input, access_token_input, access_token_secret_input)
}
}
