% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_twitter_msg.R
\name{load_twitter_by_keywords}
\alias{load_twitter_by_keywords}
\title{Load dataframe which contains specific keyword return as a dataframe with total tweets}
\usage{
load_twitter_by_keywords(keywords, limit, user_info)
}
\arguments{
\item{keywords}{A str contains keywords}

\item{limit}{An integer shows how many tweets user want to return}

\item{user_info}{A list contains user info}
}
\value{
A dataframe
}
\description{
Load dataframe which contains specific keyword return as a dataframe with total tweets
}
\examples{
\dontrun{
x <- load_twitter_by_keywords('nice', 50, user_info)
}
}
