% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_twitter_msg.R
\name{load_twitter_by_user}
\alias{load_twitter_by_user}
\title{Load dataframe which contains specific user and return as a dataframe with total tweets}
\usage{
load_twitter_by_user(keyword, limit, user_info)
}
\arguments{
\item{limit}{An integer shows how many tweets user want to return}

\item{user_info}{A list contains all user credentials}

\item{key}{A str contains tweets keyword}
}
\value{
A dataframe contains all tweets contains key words you passed
}
\description{
Load dataframe which contains specific user and return as a dataframe with total tweets
}
\examples{
\dontrun{
x <- load_twitter_by_keywords('2022', 100, user_info)
}
}
